elibrary/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
cloudinary==1.22.0
cryptography==3.1.1
defusedxml==0.6.0
dj-stripe==2.3.0
Django==2.2.10
django-appconf==1.0.4
django-cloudinary-storage==0.3.0
django-crispy-forms==1.9.2
django-downloadview==2.1.1
django-extensions==2.2.5
django-fsm==2.7.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-ratelimit==3.0.1
django-storages==1.10.1
django-tinymce4-lite==1.8.0
docutils==0.16
gunicorn==20.0.4
idna==2.8
jsmin==2.2.2
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dotenv==0.14.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sorl-thumbnail==12.5.0
sqlparse==0.3.0
stripe==2.49.0
urllib3==1.25.7
Werkzeug==0.16.0
